cr-256: refactor structure_specific_reader_aux name
