+++ src/pysdmx/io/xml/__ss_aux_reader.py
